requirements.txt
torch>=2.0.0
datasets>=2.14.0
tokenizers>=0.14.0
tqdm>=4.66.0
mlflow>=2.10.0
tensorboard>=2.20.0
Flask>=3.0.0
apache-airflow>=2.8.0
apache-airflow-providers-docker>=3.9.0
